debate_dragon/requirements.txt
discord.py==1.7.3
discord==1.7.3
Pillow==8.4.0
python-dotenv==0.19.2
